requirements.txt
pyserial==3.5
pyinstaller==5.13.0